refactor: colocate GLM speculative kernels
